Establish four functions
Created event listeners for number and operator buttons that would
allow user to type into the display and then operate on the number
in the main portion of the display. Filled out the history div to
show what number and operation are stored for the next execution when
an operator button is hit. Noticed issue with input being unlimited and
causing the display to grow too wide but have not implemented a fix
yet. Observed similar issue with digits after the decimal point that
will need to be solved with some form of rounding function to prevent
calculator from growing too wide on the page.