Polish calculator functions
Added logic to js file to handle equals, clear, and clear all buttons.
Added rounding to math functions to limit to two decimal places. Will
likely return to this project to add the decimal operations and limit
the length of numeric inputs and outputs, but feel I have grasped all
of the techniques being tested with this project, and feel confident
to move forward into the next sections to learn additional methods and
begin development of my end goal... a full website for my sports
analytics.